Partially completed switch to Tesla API.
This code was able to do an OAuth exchange with Tesla's API,
retreiving an Access Token and Refresh Token.

Tesla implements the behavior where the access token is short
lived and any use of the refresh token invalidates it while
supplying a new Refresh Token. That part isn't working yet.

I'd planned to split this into an OAuth component running
in Fly and a Prometheus component running locally, which this
code structure reflects. sherwood-energy-mon works, the
powerwall directory is only partially working in this commit.

Signed-off-by: Denton Gentry <dgentry@decarbon.earth>